+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 **Features**
 - Added `ignoreTransparency` option (see [#4](https://github.com/justintaddei/v-shared-element/issues/4) for why)
+- Added `sharedElementMixin` for `keep-alive routes` to resolve [#2](https://github.com/justintaddei/v-shared-element/issues/2)
 - Updated *illusory* to v1.4.1
 
 ---